functions/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.7.0
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
exceptiongroup==1.1.1
Flask==2.3.2
Flask-Cors==3.0.10
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.7
openai-async==0.0.3
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.6.2
urllib3==2.0.2
Werkzeug==2.3.4
yarl==1.9.2
zipp==3.15.0
